+++ custom_components/veolia/coordinator.py
@@ -6,8 +6,6 @@ from datetime import date, datetime, timedelta
 from typing import TYPE_CHECKING
 
 from dateutil.relativedelta import relativedelta
-from veolia_api import VeoliaAPI
-from veolia_api.exceptions import VeoliaAPIError
 
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.exceptions import ConfigEntryAuthFailed
@@ -18,6 +16,8 @@ from homeassistant.util import dt as dt_util
 from .const import CONF_PORTAL_URL, DOMAIN, LOGGER
 from .data import VeoliaConfigEntry
 from .model import VeoliaModel
+from .veolia_api import VeoliaAPI
+from .veolia_api.exceptions import VeoliaAPIError
 
 if TYPE_CHECKING:
     from homeassistant.core import HomeAssistant