+++ custom_components/veolia/data.py
@@ -8,11 +8,10 @@ from typing import TYPE_CHECKING
 from homeassistant.config_entries import ConfigEntry
 
 if TYPE_CHECKING:
-    from veolia_api import VeoliaAPI
-
     from homeassistant.loader import Integration
 
     from .coordinator import VeoliaDataUpdateCoordinator
+    from .veolia_api import VeoliaAPI
 
 type VeoliaConfigEntry = ConfigEntry[VeoliaData]